requirements.txt
openai
langchain
streamlit
python-dotenv
watchdog  # to supress warning
pre-commit
plotly
